Get a dataset from the Open Government Data of Thailand API.

Get a dataset from the Open Government Data of Thailand API.

Usage

th_get(resource_id)

Arguments

resource_id

Resource ID of the dataset to download.

Examples

th_get("245c98c3-2f92-4d58-a892-3fabf2af0772")
#>      _id          crash_date crash_time             expw_step         cause
#>   1:   1 2015-10-01T00:00:00   09:25:25                บูรพาวิถี          คลัทช์
#>   2:   2 2015-10-01T00:00:00   20:10:10            เฉลิมมหานคร        ยางแตก
#>   3:   3 2015-10-01T00:00:00   20:55:55            เฉลิมมหานคร        ยางแตก
#>   4:   4 2015-10-01T00:00:00   10:00:00            เฉลิมมหานคร        ยางแตก
#>   5:   5 2015-10-01T00:00:00   23:40:40                ฉลองรัช        ยางแตก
#>   6:   6 2015-10-01T00:00:00   14:19:19                ฉลองรัช        ยางแตก
#>   7:   7 2015-10-01T00:00:00   13:23:23                บูรพาวิถี        ยางแตก
#>   8:   8 2015-10-01T00:00:00   11:03:03                ฉลองรัช        ยางแตก
#>   9:   9 2015-10-01T00:00:00   08:10:10                ฉลองรัช        ยางแตก
#>  10:  10 2015-10-01T00:00:00   23:20:20                ฉลองรัช        ยางแตก
#>  11:  11 2015-10-01T00:00:00   08:22:22                ฉลองรัช        ยางแตก
#>  12:  12 2015-10-01T00:00:00   13:27:27                บูรพาวิถี        ยางแตก
#>  13:  13 2015-10-01T00:00:00   09:24:24                ฉลองรัช        ยางแตก
#>  14:  14 2015-10-01T00:00:00   08:58:58                บูรพาวิถี        ยางแตก
#>  15:  15 2015-10-01T00:00:00   06:26:26                บูรพาวิถี        ยางแตก
#>  16:  16 2015-10-01T00:00:00   08:31:31                บูรพาวิถี        ยางแตก
#>  17:  17 2015-10-01T00:00:00   21:37:37          บางพลี-สุขสวัสดิ์        ยางแตก
#>  18:  18 2015-10-01T00:00:00   17:50:50          บางพลี-สุขสวัสดิ์        ยางแตก
#>  19:  19 2015-10-01T00:00:00   16:49:49          บางพลี-สุขสวัสดิ์        ยางแตก
#>  20:  20 2015-10-01T00:00:00   14:43:43          บางพลี-สุขสวัสดิ์        ยางแตก
#>  21:  21 2015-10-01T00:00:00   16:51:51          บางพลี-สุขสวัสดิ์        ยางแตก
#>  22:  22 2015-10-01T00:00:00   21:06:06 ทางหลวงพิเศษหมายเลข 37        ยางแตก
#>  23:  23 2015-10-01T00:00:00   16:00:00          บางพลี-สุขสวัสดิ์        ยางแตก
#>  24:  24 2015-10-01T00:00:00   15:51:51 ทางหลวงพิเศษหมายเลข 37        ยางแตก
#>  25:  25 2015-10-01T00:00:00   15:45:45 ทางหลวงพิเศษหมายเลข 37        ยางแตก
#>  26:  26 2015-10-01T00:00:00   08:02:02          บางพลี-สุขสวัสดิ์        ยางแตก
#>  27:  27 2015-10-01T00:00:00   06:53:53 ทางหลวงพิเศษหมายเลข 37        ยางแตก
#>  28: 135 2015-10-02T00:00:00   18:42:42                  ศรีรัช      ระบบเบรค
#>  29:  28 2015-10-01T00:00:00   03:15:15 ทางหลวงพิเศษหมายเลข 37        ยางแตก
#>  30:  29 2015-10-01T00:00:00   08:51:51               อุดรรัถยา        ยางแตก
#>  31:  30 2015-10-01T00:00:00   11:04:04               อุดรรัถยา        ยางแตก
#>  32:  31 2015-10-01T00:00:00   06:25:25            เฉลิมมหานคร        ยางแตก
#>  33:  32 2015-10-01T00:00:00   08:24:24                  ศรีรัช        ยางแตก
#>  34:  33 2015-10-01T00:00:00   13:50:50            เฉลิมมหานคร        ยางแตก
#>  35:  34 2015-10-01T00:00:00   23:06:06          บางพลี-สุขสวัสดิ์        ยางแตก
#>  36:  35 2015-10-01T00:00:00   18:10:10          บางพลี-สุขสวัสดิ์        ยางแตก
#>  37:  36 2015-10-01T00:00:00   14:53:53                  ศรีรัช        ยางแตก
#>  38:  37 2015-10-01T00:00:00   17:29:29                  ศรีรัช        ยางแตก
#>  39:  38 2015-10-01T00:00:00   07:54:54            เฉลิมมหานคร        ระบบไฟ
#>  40:  39 2015-10-01T00:00:00   23:28:28            เฉลิมมหานคร        ระบบไฟ
#>  41:  40 2015-10-01T00:00:00   02:21:21 ทางหลวงพิเศษหมายเลข 37        ระบบไฟ
#>  42:  41 2015-10-01T00:00:00   07:58:58            เฉลิมมหานคร         น้ำแห้ง
#>  43:  42 2015-10-01T00:00:00   18:13:13                บูรพาวิถี         น้ำแห้ง
#>  44:  43 2015-10-01T00:00:00   17:39:39          บางพลี-สุขสวัสดิ์         น้ำแห้ง
#>  45:  44 2015-10-01T00:00:00   15:40:40          บางพลี-สุขสวัสดิ์         น้ำแห้ง
#>  46:  45 2015-10-01T00:00:00   16:00:00            เฉลิมมหานคร         น้ำแห้ง
#>  47:  46 2015-10-01T00:00:00   23:22:22            เฉลิมมหานคร       ตกร่องน้ำ
#>  48:  47 2015-10-01T00:00:00   20:00:00            เฉลิมมหานคร       น้ำมันหมด
#>  49:  48 2015-10-01T00:00:00   19:19:19                ฉลองรัช       น้ำมันหมด
#>  50:  49 2015-10-01T00:00:00   14:58:58                ฉลองรัช       น้ำมันหมด
#>  51:  50 2015-10-01T00:00:00   15:23:23                บูรพาวิถี       น้ำมันหมด
#>  52:  51 2015-10-01T00:00:00   16:45:45                บูรพาวิถี       น้ำมันหมด
#>  53:  52 2015-10-01T00:00:00   23:02:02 ทางหลวงพิเศษหมายเลข 37       น้ำมันหมด
#>  54:  53 2015-10-01T00:00:00   19:33:33          บางพลี-สุขสวัสดิ์       น้ำมันหมด
#>  55:  54 2015-10-01T00:00:00   16:30:30            เฉลิมมหานคร       น้ำมันหมด
#>  56:  55 2015-10-01T00:00:00   16:51:51                ฉลองรัช      ระบบคลัทช์
#>  57:  56 2015-10-01T00:00:00   10:59:59          บางพลี-สุขสวัสดิ์      ระบบคลัทช์
#>  58:  57 2015-10-01T00:00:00   16:44:44                  ศรีรัช      เกียร์ชำรุด
#>  59:  58 2015-10-01T00:00:00   10:02:02                ฉลองรัช    ระบบส่งกำลัง
#>  60:  59 2015-10-01T00:00:00   09:32:32          บางพลี-สุขสวัสดิ์  เพลากลางชำรุด
#>  61:  60 2015-10-01T00:00:00   17:30:30            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  62:  61 2015-10-01T00:00:00   09:45:45            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  63:  62 2015-10-01T00:00:00   17:33:33                ฉลองรัช เครื่องยนต์ขัดข้อง
#>  64:  63 2015-10-01T00:00:00   08:07:07                ฉลองรัช เครื่องยนต์ขัดข้อง
#>  65:  64 2015-10-01T00:00:00   16:56:56                ฉลองรัช เครื่องยนต์ขัดข้อง
#>  66:  65 2015-10-01T00:00:00   14:07:07                บูรพาวิถี เครื่องยนต์ขัดข้อง
#>  67:  66 2015-10-01T00:00:00   11:21:21                บูรพาวิถี เครื่องยนต์ขัดข้อง
#>  68:  67 2015-10-01T00:00:00   08:35:35                บูรพาวิถี เครื่องยนต์ขัดข้อง
#>  69:  68 2015-10-01T00:00:00   08:30:30                บูรพาวิถี เครื่องยนต์ขัดข้อง
#>  70:  69 2015-10-01T00:00:00   07:29:29                บูรพาวิถี เครื่องยนต์ขัดข้อง
#>  71:  70 2015-10-01T00:00:00   07:00:00          บางพลี-สุขสวัสดิ์ เครื่องยนต์ขัดข้อง
#>  72:  71 2015-10-01T00:00:00   16:02:02          บางพลี-สุขสวัสดิ์ เครื่องยนต์ขัดข้อง
#>  73:  72 2015-10-01T00:00:00   19:43:43          บางพลี-สุขสวัสดิ์ เครื่องยนต์ขัดข้อง
#>  74:  73 2015-10-01T00:00:00   16:43:43          บางพลี-สุขสวัสดิ์ เครื่องยนต์ขัดข้อง
#>  75:  74 2015-10-01T00:00:00   07:16:16          บางพลี-สุขสวัสดิ์ เครื่องยนต์ขัดข้อง
#>  76:  75 2015-10-01T00:00:00   09:47:47               อุดรรัถยา เครื่องยนต์ขัดข้อง
#>  77:  76 2015-10-01T00:00:00   16:42:42                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  78:  77 2015-10-01T00:00:00   16:38:38            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  79:  78 2015-10-01T00:00:00   20:46:46                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  80:  79 2015-10-01T00:00:00   06:36:36                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  81:  80 2015-10-01T00:00:00   10:39:39            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  82:  81 2015-10-01T00:00:00   16:00:00            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  83:  82 2015-10-01T00:00:00   21:50:50            เฉลิมมหานคร เครื่องยนต์ขัดข้อง
#>  84:  83 2015-10-01T00:00:00   11:06:06                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  85:  84 2015-10-01T00:00:00   08:52:52                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  86:  85 2015-10-01T00:00:00   08:15:15                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  87:  86 2015-10-01T00:00:00   19:11:11                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  88:  87 2015-10-01T00:00:00   17:30:30                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  89:  88 2015-10-01T00:00:00   17:14:14                  ศรีรัช เครื่องยนต์ขัดข้อง
#>  90:  89 2015-10-02T00:00:00   18:34:34            เฉลิมมหานคร        ยางแตก
#>  91:  90 2015-10-02T00:00:00   07:00:00            เฉลิมมหานคร        ยางแตก
#>  92:  91 2015-10-02T00:00:00   07:28:28            เฉลิมมหานคร        ยางแตก
#>  93:  92 2015-10-02T00:00:00   17:42:42            เฉลิมมหานคร        ยางแตก
#>  94:  93 2015-10-02T00:00:00   11:52:52            เฉลิมมหานคร        ยางแตก
#>  95:  94 2015-10-02T00:00:00   19:04:04                ฉลองรัช        ยางแตก
#>  96:  95 2015-10-02T00:00:00   08:15:15                ฉลองรัช        ยางแตก
#>  97:  96 2015-10-02T00:00:00   15:38:38                บูรพาวิถี        ยางแตก
#>  98:  97 2015-10-02T00:00:00   13:32:32                บูรพาวิถี        ยางแตก
#>  99:  98 2015-10-02T00:00:00   12:41:41                บูรพาวิถี        ยางแตก
#> 100:  99 2015-10-02T00:00:00   17:17:17                บูรพาวิถี        ยางแตก
#>      _id          crash_date crash_time             expw_step         cause